backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
tensorflow==2.20.0
numpy>=1.26.0
Pillow>=10.0.0
gunicorn>=21.0.0
transformers>=4.45.0
tokenizers>=0.20.3
tf_keras>=2.20.1
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
gradio==4.44.0